@@ -95,7 +95,7 @@
     "copy-webpack-plugin": "13.0.0",
     "cross-env": "10.0.0",
     "css-loader": "7.1.2",
-    "electron": "36.8.1",
+    "electron": "38.2.0",
     "electron-builder": "26.0.12",
     "electron-log": "5.4.0",
     "electron-reload": "2.0.0-alpha.1",
@@ -158,7 +158,7 @@
     "@angular/platform-browser": "19.2.14",
     "@angular/platform-browser-dynamic": "19.2.14",
     "@angular/router": "19.2.14",
-    "@bitwarden/sdk-internal": "0.2.0-main.296",
+    "@bitwarden/sdk-internal": "0.2.0-main.311",
     "@electron/fuses": "1.8.0",
     "@emotion/css": "11.13.5",
     "@koa/multer": "4.0.0",
