@@ -1,4 +1,4 @@
-import { combineLatest, distinctUntilChanged, mergeMap, of, switchMap, withLatestFrom } from "rxjs";
+import { combineLatest, delay, distinctUntilChanged, mergeMap, of, switchMap } from "rxjs";
 
 import { AccountService } from "@bitwarden/common/auth/abstractions/account.service";
 import { BadgeSettingsServiceAbstraction } from "@bitwarden/common/autofill/services/badge-settings.service";
@@ -10,7 +10,7 @@ import { Tab } from "../../platform/badge/badge-browser-api";
 import { BadgeService } from "../../platform/badge/badge.service";
 import { BadgeStatePriority } from "../../platform/badge/priority";
 
-const StateName = (tabId: number) => `autofill-badge-${tabId}`;
+const StateName = "autofill-badge-updater";
 
 export class AutofillBadgeUpdaterService {
   constructor(
@@ -26,56 +26,30 @@ export class AutofillBadgeUpdaterService {
       switchMap((account) => (account?.id ? this.cipherService.ciphers$(account?.id) : of([]))),
     );
 
-    // Recalculate badges for all active tabs when ciphers or active account changes
-    combineLatest({
-      account: this.accountService.activeAccount$,
-      enableBadgeCounter:
-        this.badgeSettingsService.enableBadgeCounter$.pipe(distinctUntilChanged()),
-      ciphers: ciphers$,
-    })
-      .pipe(
+    this.badgeService.setState(StateName, (tab) => {
+      return combineLatest({
+        account: this.accountService.activeAccount$,
+        enableBadgeCounter:
+          this.badgeSettingsService.enableBadgeCounter$.pipe(distinctUntilChanged()),
+        ciphers: ciphers$.pipe(delay(100)), // Delay to allow cipherService.getAllDecryptedForUrl to pick up changes
+      }).pipe(
         mergeMap(async ({ account, enableBadgeCounter }) => {
-          if (!account) {
-            return;
-          }
-
-          const tabs = await this.badgeService.getActiveTabs();
-
-          for (const tab of tabs) {
-            if (!tab.tabId) {
-              continue;
-            }
-            if (enableBadgeCounter) {
-              await this.setTabState(tab, account.id);
-            } else {
-              await this.clearTabState(tab.tabId);
-            }
-          }
-        }),
-      )
-      .subscribe();
-
-    // Recalculate badge for a specific tab when it becomes active
-    this.badgeService.activeTabsUpdated$
-      .pipe(
-        withLatestFrom(
-          this.accountService.activeAccount$,
-          this.badgeSettingsService.enableBadgeCounter$,
-        ),
-        mergeMap(async ([tabs, account, enableBadgeCounter]) => {
           if (!account || !enableBadgeCounter) {
-            return;
+            return undefined;
           }
 
-          for (const tab of tabs) {
-            await this.setTabState(tab, account.id);
-          }
+          return {
+            state: {
+              text: await this.calculateCountText(tab, account.id),
+            },
+            priority: BadgeStatePriority.Default,
+          };
         }),
-      )
-      .subscribe();
+      );
+    });
   }
 
-  private async setTabState(tab: Tab, userId: UserId) {
+  private async calculateCountText(tab: Tab, userId: UserId) {
     if (!tab.tabId) {
       this.logService.warning("Tab event received but tab id is undefined");
       return;
@@ -85,22 +59,9 @@ export class AutofillBadgeUpdaterService {
     const cipherCount = ciphers.length;
 
     if (cipherCount === 0) {
-      await this.clearTabState(tab.tabId);
-      return;
+      return undefined;
     }
 
-    const countText = cipherCount > 9 ? "9+" : cipherCount.toString();
-    await this.badgeService.setState(
-      StateName(tab.tabId),
-      BadgeStatePriority.Default,
-      {
-        text: countText,
-      },
-      tab.tabId,
-    );
-  }
-
-  private async clearTabState(tabId: number) {
-    await this.badgeService.clearState(StateName(tabId));
+    return cipherCount > 9 ? "9+" : cipherCount.toString();
   }
 }
