@@ -1,11 +1,20 @@
 // FIXME: Update this file to be type safe and remove this and next line
 // @ts-strict-ignore
 import { SelectionModel } from "@angular/cdk/collections";
-import { Component, EventEmitter, Input, OnDestroy, Output } from "@angular/core";
-import { Subject, takeUntil } from "rxjs";
+import { Component, EventEmitter, Input, OnDestroy, Output, OnInit } from "@angular/core";
+import { ActivatedRoute } from "@angular/router";
+import { catchError, concatMap, map, Observable, of, Subject, switchMap, takeUntil } from "rxjs";
 
+import {
+  getOrganizationById,
+  OrganizationService,
+} from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
+import { AccountService } from "@bitwarden/common/auth/abstractions/account.service";
+import { getUserId } from "@bitwarden/common/auth/services/account.service";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
-import { TableDataSource, ToastService } from "@bitwarden/components";
+import { DialogRef, DialogService, TableDataSource, ToastService } from "@bitwarden/components";
+import { LogService } from "@bitwarden/logging";
+import { openEntityEventsDialog } from "@bitwarden/web-vault/app/admin-console/organizations/manage/entity-events.component";
 
 import {
   ServiceAccountSecretsDetailsView,
@@ -17,7 +26,7 @@ import {
   templateUrl: "./service-accounts-list.component.html",
   standalone: false,
 })
-export class ServiceAccountsListComponent implements OnDestroy {
+export class ServiceAccountsListComponent implements OnDestroy, OnInit {
   protected dataSource = new TableDataSource<ServiceAccountSecretsDetailsView>();
 
   @Input()
@@ -43,16 +52,50 @@ export class ServiceAccountsListComponent implements OnDestroy {
   @Output() editServiceAccountEvent = new EventEmitter<string>();
 
   private destroy$: Subject<void> = new Subject<void>();
-
+  protected viewEventsAllowed$: Observable<boolean>;
+  protected isAdmin$: Observable<boolean>;
   selection = new SelectionModel<string>(true, []);
 
   constructor(
     private i18nService: I18nService,
     private toastService: ToastService,
+    private dialogService: DialogService,
+    private organizationService: OrganizationService,
+    private activatedRoute: ActivatedRoute,
+    private accountService: AccountService,
+    private logService: LogService,
   ) {
     this.selection.changed
       .pipe(takeUntil(this.destroy$))
       .subscribe((_) => this.onServiceAccountCheckedEvent.emit(this.selection.selected));
+  }
+
+  ngOnInit(): void {
+    this.viewEventsAllowed$ = this.activatedRoute.params.pipe(
+      concatMap((params) =>
+        getUserId(this.accountService.activeAccount$).pipe(
+          switchMap((userId) =>
+            this.organizationService
+              .organizations$(userId)
+              .pipe(getOrganizationById(params.organizationId)),
+          ),
+        ),
+      ),
+      map((org) => org.canAccessEventLogs),
+      catchError((error: unknown) => {
+        if (typeof error === "string") {
+          this.toastService.showToast({
+            message: error,
+            variant: "error",
+            title: "",
+          });
+        } else {
+          this.logService.error(error);
+        }
+        return of(false);
+      }),
+      takeUntil(this.destroy$),
+    );
   }
 
   ngOnDestroy(): void {
@@ -94,4 +137,13 @@ export class ServiceAccountsListComponent implements OnDestroy {
       });
     }
   }
+  openEventsDialog = (serviceAccount: ServiceAccountView): DialogRef<void> =>
+    openEntityEventsDialog(this.dialogService, {
+      data: {
+        name: serviceAccount.name,
+        organizationId: serviceAccount.organizationId,
+        entityId: serviceAccount.id,
+        entity: "service-account",
+      },
+    });
 }
