@@ -4,7 +4,7 @@ import { Directive, EventEmitter, Input, Output } from "@angular/core";
 
 // This import has been flagged as unallowed for this class. It may be involved in a circular dependency loop.
 // eslint-disable-next-line no-restricted-imports
-import { CollectionView } from "@bitwarden/admin-console/common";
+import { CollectionTypes, CollectionView } from "@bitwarden/admin-console/common";
 import { ITreeNodeObject } from "@bitwarden/common/vault/models/domain/tree-node";
 
 import { DynamicTreeNode } from "../models/dynamic-tree-node.model";
@@ -21,6 +21,7 @@ export class CollectionFilterComponent {
   @Output() onNodeCollapseStateChange: EventEmitter<ITreeNodeObject> =
     new EventEmitter<ITreeNodeObject>();
   @Output() onFilterChange: EventEmitter<VaultFilter> = new EventEmitter<VaultFilter>();
+  DefaultCollectionType = CollectionTypes.DefaultUserCollection;
 
   readonly collectionsGrouping: TopLevelTreeNode = {
     id: "collections",
