@@ -325,11 +325,11 @@ export class CipherView implements View, InitializerMetadata {
       name: this.name ?? "",
       notes: this.notes,
       type: this.type ?? CipherType.Login,
-      favorite: this.favorite,
-      organizationUseTotp: this.organizationUseTotp,
+      favorite: this.favorite ?? false,
+      organizationUseTotp: this.organizationUseTotp ?? false,
       permissions: this.permissions?.toSdkCipherPermissions(),
-      edit: this.edit,
-      viewPassword: this.viewPassword,
+      edit: this.edit ?? true,
+      viewPassword: this.viewPassword ?? true,
       localData: toSdkLocalData(this.localData),
       attachments: this.attachments?.map((a) => a.toSdkAttachmentView()),
       fields: this.fields?.map((f) => f.toSdkFieldView()),
