@@ -5,13 +5,14 @@ import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.servic
 import { SharedModule } from "../shared";
 import { TypographyModule } from "../typography";
 
-export type CalloutTypes = "success" | "info" | "warning" | "danger";
+export type CalloutTypes = "success" | "info" | "warning" | "danger" | "default";
 
 const defaultIcon: Record<CalloutTypes, string> = {
   success: "bwi-check-circle",
   info: "bwi-info-circle",
   warning: "bwi-exclamation-triangle",
   danger: "bwi-error",
+  default: "bwi-star",
 };
 
 const defaultI18n: Partial<Record<CalloutTypes, string>> = {
@@ -55,13 +56,15 @@ export class CalloutComponent {
   protected readonly calloutClass = computed(() => {
     switch (this.type()) {
       case "danger":
-        return "tw-bg-danger-100";
+        return "tw-bg-danger-100 tw-border-danger-700 tw-text-danger-700";
       case "info":
-        return "tw-bg-info-100";
+        return "tw-bg-info-100 tw-bg-info-100 tw-border-info-700 tw-text-info-700";
       case "success":
-        return "tw-bg-success-100";
+        return "tw-bg-success-100 tw-bg-success-100 tw-border-success-700 tw-text-success-700";
       case "warning":
-        return "tw-bg-warning-100";
+        return "tw-bg-warning-100 tw-bg-warning-100 tw-border-warning-700 tw-text-warning-700";
+      case "default":
+        return "tw-bg-background-alt tw-border-secondary-700 tw-text-secondary-700";
     }
   });
 }
