@@ -5,12 +5,10 @@ import { Component, effect, input } from "@angular/core";
 import { firstValueFrom, map } from "rxjs";
 
 import { JslibModule } from "@bitwarden/angular/jslib.module";
-import {
-  getOrganizationById,
-  OrganizationService,
-} from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
+import { OrganizationService } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { AccountService } from "@bitwarden/common/auth/abstractions/account.service";
 import { getUserId } from "@bitwarden/common/auth/services/account.service";
+import { getById } from "@bitwarden/common/platform/misc/rxjs-operators";
 import { CalloutModule } from "@bitwarden/components";
 
 @Component({
@@ -30,6 +28,8 @@ export class ExportScopeCalloutComponent {
   readonly organizationId = input<string>();
   /* Optional export format, determines which individual export description to display */
   readonly exportFormat = input<string>();
+  /* The description key to use for organizational exports */
+  readonly orgExportDescription = input<string>();
 
   constructor(
     protected organizationService: OrganizationService,
@@ -37,35 +37,45 @@ export class ExportScopeCalloutComponent {
   ) {
     effect(async () => {
       this.show = false;
-      await this.getScopeMessage(this.organizationId(), this.exportFormat());
+      await this.getScopeMessage(
+        this.organizationId(),
+        this.exportFormat(),
+        this.orgExportDescription(),
+      );
       this.show = true;
     });
   }
 
-  private async getScopeMessage(organizationId: string, exportFormat: string): Promise<void> {
+  private async getScopeMessage(
+    organizationId: string,
+    exportFormat: string,
+    orgExportDescription: string,
+  ): Promise<void> {
     const userId = await firstValueFrom(getUserId(this.accountService.activeAccount$));
-    this.scopeConfig =
-      organizationId != null
-        ? {
-            title: "exportingOrganizationVaultTitle",
-            description: "exportingOrganizationVaultDesc",
-            scopeIdentifier: (
-              await firstValueFrom(
-                this.organizationService
-                  .organizations$(userId)
-                  .pipe(getOrganizationById(organizationId)),
-              )
-            ).name,
-          }
-        : {
-            title: "exportingPersonalVaultTitle",
-            description:
-              exportFormat == "zip"
-                ? "exportingIndividualVaultWithAttachmentsDescription"
-                : "exportingIndividualVaultDescription",
-            scopeIdentifier: await firstValueFrom(
-              this.accountService.activeAccount$.pipe(map((a) => a?.email)),
-            ),
-          };
+
+    if (organizationId != null) {
+      // exporting from organizational vault
+      const org = await firstValueFrom(
+        this.organizationService.organizations$(userId).pipe(getById(organizationId)),
+      );
+
+      this.scopeConfig = {
+        title: "exportingOrganizationVaultTitle",
+        description: orgExportDescription,
+        scopeIdentifier: org?.name ?? "",
+      };
+    } else {
+      this.scopeConfig = {
+        // exporting from individual vault
+        title: "exportingPersonalVaultTitle",
+        description:
+          exportFormat === "zip"
+            ? "exportingIndividualVaultWithAttachmentsDescription"
+            : "exportingIndividualVaultDescription",
+        scopeIdentifier:
+          (await firstValueFrom(this.accountService.activeAccount$.pipe(map((a) => a?.email)))) ??
+          "",
+      };
+    }
   }
 }
